Advance original build declaration frontier
Expand the non-Windows Win32 declaration surface enough for WORD1 and opus_word1_ui_test to compile on macOS.

Gate Windows import libraries to WIN32, feed the original include directories into the product/test targets, and use a Darwin dynamic-loader handoff so missing shim bodies become runtime failures instead of declaration blockers.

Fix the spell/file/startup/chrome/UI harness compile frontiers with pointer-sized casts, callback typing, local SDK forward headers, and narrow portability fixes.

Validation: commentflow on changed shim/original headers and focused spell/prompt/RTF files; git diff --check; cmake --build out/macos-debug --target opus_word1_ui_test. Smoke test now aborts at missing _SymFunctionTableAccess64, the next DbgHelp shim body.

+++ src/port/original/opus_word1_ui_test.cpp
@@ -220,7 +220,7 @@ std::size_t count_dark_client_pixels(const HWND window,
     if (window == nullptr || !GetClientRect(window, &client)) {
         return 0;
     }
-    const int width = (std::min)(client.right, 800L);
+    const int width = (std::min)(client.right, static_cast<LONG>(800));
     const int top = (std::max)(0, y_first);
     const int bottom = (std::min)(client.bottom, static_cast<LONG>(y_limit));
     const int height = bottom - top;
@@ -431,8 +431,8 @@ bool make_foreground_and_focus(const HWND main_window, const HWND focus,
     if (!GetClientRect(focus, &client)) {
         return false;
     }
-    activation_point.x = (std::min)(client.right - 1, 20L);
-    activation_point.y = (std::min)(client.bottom - 1, 10L);
+    activation_point.x = (std::min)(client.right - 1, static_cast<LONG>(20));
+    activation_point.y = (std::min)(client.bottom - 1, static_cast<LONG>(10));
     if (activation_point.x < 0 || activation_point.y < 0 ||
         !ClientToScreen(focus, &activation_point) ||
         !SetCursorPos(activation_point.x, activation_point.y)) {
